move ingorer to a factory
